+++ modules/Shmuelie.Git/Shmuelie.Git.psm1
@@ -72,6 +72,7 @@ Export-ModuleMember -Function @(
     'Set-Branch',
     'Format-GitStatusSegment',
     'Get-Worktrees',
+    'Get-Branch',
     'Get-CurrentWorktree',
     'Get-RepositoryName',
     'Get-RootWorktree',